backend/requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.1.8
Flask==3.1.2
flask-cors==6.0.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.46
tomli==2.4.0
typing_extensions==4.15.0
Werkzeug==3.1.5
zipp==3.23.0
